AddView (DetailViewController) -> AddContentView
