+++ RandomPicker/RandomPicker/ViewModel/MainViewModel.swift
@@ -3,7 +3,7 @@ import RxDataSources
 
 typealias ContentSectionModel = AnimatableSectionModel<Int, Content>
 
-class MainViewModel: ViewModelType {
+final class MainViewModel: ViewModelType {
     let dataSource: RxTableViewSectionedAnimatedDataSource<ContentSectionModel> = {
         let ds = RxTableViewSectionedAnimatedDataSource<ContentSectionModel> { ds, tableView, indexPath, content in
             let cell = tableView.dequeueReusableCell(withIdentifier: MainTableViewCell.identifier, for: indexPath) as! MainTableViewCell
